V1 de login et inscription

+++ src/sharemycar/app/src/main/java/com/example/sharemycar/data/models/User.kt
@@ -3,7 +3,7 @@ package com.example.sharemycar.data.models
 import com.google.gson.annotations.SerializedName
 
 class User(
-    @SerializedName("id") val idController: Int,
+    @SerializedName("id") val idController: Long,
     @SerializedName("username") val username: String,
     @SerializedName("email") val email: String,
